# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory Kamples API",
    "description": "API Kamples — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/colecciones/{id}/descargar-zip": {
      "post": {
        "tags": [
          "colecciones"
        ],
        "operationId": "descargar_zip_coleccion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la coleccion a descargar como ZIP",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DownloadGrantRequest"
                  }
                ],
                "nullable": true
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Stream binario application/zip"
          },
          "400": {
            "description": "Coleccion vacia o demasiado grande"
          },
          "403": {
            "description": "Sin acceso o samples premium en plan free"
          },
          "404": {
            "description": "Coleccion no encontrada"
          },
          "429": {
            "description": "Creditos insuficientes para los samples nuevos"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadGrantRequest": {
        "type": "object",
        "properties": {
          "codigoGratis": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}